+++ sdk/storage/Azure.Storage.Blobs/CHANGELOG.md
@@ -8,7 +8,8 @@
 ### Breaking Changes
 
 ### Bugs Fixed
-
+- Fixed bug where a `BlobServiceClient`, `BlobContainerClient`, `BlobBaseClient` created with a connection string with an account name specified (e.g. "AccountName=..;"), the account name was not populated on the Storage Clients if the account name was not also specified in the endpoint. (#42925)
+- Fixed bug where a `BlobServiceClient`, `BlobContainerClient`, `BlobBaseClient` created with a `StorageSharedKeyCredential`, the account name was not populated on the Storage Clients if the account name was not also specified in the endpoint. (#42925)
 ### Other Changes
 
 ## 12.24.0-beta.1 (2025-02-11)